cscan_iot/requirements.txt
fastapi
uvicorn
numpy
matplotlib
pyserial
python-libtiepie
scipy
jinja2
python-multipart